requirements.txt
# Base requirements
numpy>=1.20.0
annotated-types
anyio
asgiref
beautifulsoup4
boto3
botocore
certifi
chardet
charset-normalizer
contourpy
cycler
decorator
distro
dj-static
Django
django-bootstrap-v5
django-htmx
django-crontab
django-smtp-ssl
django-storages
elevenlabs
exceptiongroup
fonttools
gunicorn
h11
httpcore
httpx
idna
imageio
imageio-ffmpeg
jiter
jmespath
kiwisolver
lxml
matplotlib
moviepy
openai==1.76.0
opencv-python
packaging
pathlib
Pillow
proglog
psycopg2-binary
pydantic
pydantic_core
pyparsing
pysrt
python-dateutil
python-dotenv
requests>=2.28.0
s3transfer
six
sniffio
soupsieve
sqlparse
static-ranges
static3
stripe
tqdm>=4.64.0
typing_extensions
urllib3
websockets
apify-client
cloudconvert
modal
pilmoji
emoji
regex
aeneas>=1.7.3
celery
redis
django-redis
dj-database-url
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pytz
# GPU and Video Processing Support
ffmpeg-python
psutil
django-storages 
boto3